# Conflicts:
#	composer.json
#	composer.lock

+++ database/migrations/2018_01_01_000000_create_permission_tables.php
@@ -16,10 +16,16 @@ class CreatePermissionTables extends Migration
         $tableNames = config('permission.table_names');
         $columnNames = config('permission.column_names');
 
+        if (empty($tableNames)) {
+            throw new Exception(
+                'Error: config/permission.php not loaded. Run [php artisan config:clear] and try again.'
+            );
+        }
+
         Schema::create(
             $tableNames['permissions'],
             function (Blueprint $table) {
-                $table->increments('id');
+                $table->bigIncrements('id');
                 $table->string('name');
                 $table->string('guard_name');
                 $table->timestamps();
@@ -29,7 +35,7 @@ class CreatePermissionTables extends Migration
         Schema::create(
             $tableNames['roles'],
             function (Blueprint $table) {
-                $table->increments('id');
+                $table->bigIncrements('id');
                 $table->string('name');
                 $table->string('guard_name');
                 $table->timestamps();
@@ -39,12 +45,12 @@ class CreatePermissionTables extends Migration
         Schema::create(
             $tableNames['model_has_permissions'],
             function (Blueprint $table) use ($tableNames, $columnNames) {
-                $table->unsignedInteger('permission_id');
+                $table->unsignedBigInteger('permission_id');
 
                 $table->string('model_type');
                 $table->unsignedBigInteger($columnNames['model_morph_key']);
                 $table->index(
-                    [$columnNames['model_morph_key'], 'model_type',],
+                    [$columnNames['model_morph_key'], 'model_type'],
                     'model_has_permissions_model_id_model_type_index'
                 );
 
@@ -63,12 +69,12 @@ class CreatePermissionTables extends Migration
         Schema::create(
             $tableNames['model_has_roles'],
             function (Blueprint $table) use ($tableNames, $columnNames) {
-                $table->unsignedInteger('role_id');
+                $table->unsignedBigInteger('role_id');
 
                 $table->string('model_type');
                 $table->unsignedBigInteger($columnNames['model_morph_key']);
                 $table->index(
-                    [$columnNames['model_morph_key'], 'model_type',],
+                    [$columnNames['model_morph_key'], 'model_type'],
                     'model_has_roles_model_id_model_type_index'
                 );
 
@@ -87,8 +93,8 @@ class CreatePermissionTables extends Migration
         Schema::create(
             $tableNames['role_has_permissions'],
             function (Blueprint $table) use ($tableNames) {
-                $table->unsignedInteger('permission_id');
-                $table->unsignedInteger('role_id');
+                $table->unsignedBigInteger('permission_id');
+                $table->unsignedBigInteger('role_id');
 
                 $table->foreign('permission_id')
                     ->references('id')
@@ -117,6 +123,12 @@ class CreatePermissionTables extends Migration
     public function down()
     {
         $tableNames = config('permission.table_names');
+
+        if (empty($tableNames)) {
+            throw new Exception(
+                'Error: config/permission.php not found and defaults could not be merged. Please publish the package configuration before proceeding, or drop the tables manually.'
+            );
+        }
 
         Schema::drop($tableNames['role_has_permissions']);
         Schema::drop($tableNames['model_has_roles']);
